utils: Move ProcessCapture from core to common
The class is not specific to ORT and can be useful for other modules.
This requires to create a dedicated logger instance, because the ORT
log helper is not available in common-utils.
