+++ utils/core/src/test/kotlin/ExtensionsTest.kt
@@ -35,6 +35,7 @@ import java.io.IOException
 import java.util.Locale
 
 import org.ossreviewtoolkit.utils.common.Os
+import org.ossreviewtoolkit.utils.common.ProcessCapture
 import org.ossreviewtoolkit.utils.test.containExactly
 import org.ossreviewtoolkit.utils.test.createSpecTempDir
 import org.ossreviewtoolkit.utils.test.createTestTempDir
